Add remainder arguments
Works the same as in Discord.NET, allowing you to accept what would be
multiple arguments for just one argument of the command. I'm not much of
a fan of its implementation, but sadly I don't think there's a better
way of doing it until custom attributes or something like that happens.