Add command_service and reorganize project
A LOT of breaking changes in this one:
- Header files moved around a lot. You will almost certainly have to
  change your #includes a bit.
- module_service_config renamed to command_service_config
- function_traits moved to dpp::utility namespace

Next up, the new command_service. This allows you to take advantage of
the conveniences that the command handler provides without having to
create module classes and all that stuff. Most of the reorganizing was
done to share as much between this and module_service as possible.